[fix & refactor] user 승하차지 변경 usecase수정
이름을 더 직관적인 changeuUserStation으로 변경,
dto에 roadNameAddress필드 추가
