+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/user/exchanges/UpdateHomeLocationDto.java
@@ -4,9 +4,8 @@ import lombok.Getter;
 
 @Getter
 public class UpdateHomeLocationDto {
-    private Double houseLatitude;
-    private Double houseLongitude;
-    private Double stationLatitude;
-    private Double stationLongitude;
     private String stationName;
+    private String stationRoadAddressName;
+    private Double latitude;
+    private Double longitude;
 }
